requirements.txt
loguru==0.7.3
PyPDF2==3.0.1
pinecone[grpc]
pinecone-client
langchain-pinecone==0.2.0
streamlit
langchain-community 
langgraph 
langchain-anthropic 
tavily-python 
langgraph-checkpoint-sqlite
langchain-openai
litellm
langchain-mistralai